+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version "next"
 * Feature: Implement a serial version of LOMP runtime.
 * Feature: Support static-library version of LOMP
+* Internal: use aligned allocations instead of regular malloc/free
 
 # Version 0.2
 * Feature: Add entrypoints for tasks with `if(0)` clause.